% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADANetworks-package.R
\docType{package}
\name{ADANetworks-package}
\alias{ADANetworks}
\alias{ADANetworks-package}
\title{ADANetworks: Materials and methods for network analysis. Created for Applied Data Analysis class, Spring 2024.}
\description{
The functions and materials included in this package are designed to give a brief introduction to the construction, quantification, and analysis of ecological networks using the {bipartite} package.
}
\keyword{internal}
