% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{network}
\alias{network}
\title{Basic visualization & network-level analysis of interaction webs}
\arguments{
\item{d}{species interaction matrix data set}

\item{alternative}{network metric to produce, defaults to "all"}

\item{plot}{produce network visualization, defaults to TRUE}
}
\description{
Calculates five major network indices and visualizes network interactions.
}
\examples{
network(d = NULL, alternative = "all", plot = TRUE)
}
\keyword{networks}
