% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{null_network}
\alias{null_network}
\title{Generate & visualize an equivalent null model for your interaction matrix}
\usage{
null_network(d, alternative = "all", plot = TRUE)
}
\arguments{
\item{d}{species interaction matrix data set}

\item{alternative}{network metric to produce, defaults to "all"}

\item{plot}{produce network visualization, defaults to TRUE}
}
\description{
Generates, analyzes, and visualizes a random (null-model) matrix based on the number of higher- and lower-level species and the average number of node-to-node links in the original matrix.
}
\examples{
null_network(d = NULL, alternative = "all", plot = TRUE)
}
\keyword{networks}
