% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{example_data}
\alias{example_data}
\title{Import example interaction matrix}
\arguments{
\item{x}{matrix of plant-bird interactions}

\item{abbrev_names}{abbreviates names for clearer display in network plots, defaults to FALSE
example_data(x=NULL, abrev_names= FALSE)}
}
\description{
Imports a matrix of plant-bird interactions that can be saved to the user environment
}
